+++ App.py
@@ -12,14 +12,11 @@ from json_utils import convert_deques
 import signal
 import sys
 import threading
-import requests
-import csv
-import io
-from flask import Flask, Response, jsonify, render_template, request
+from flask import Flask, jsonify, render_template, request
 from datetime import datetime, timedelta
 from zoneinfo import ZoneInfo
 from flask_caching import Cache
-from notification_service import NotificationLevel, NotificationCategory
+from notification_service import NotificationCategory
 from urllib.parse import urlparse
 from data_service import MiningDashboardService  # noqa: F401 - used in tests
 from apscheduler.schedulers.background import BackgroundScheduler  # noqa: F401 - used in tests
@@ -55,6 +52,7 @@ import notification_routes
 import scheduler_service
 import config_routes
 import memory_routes
+import earnings_routes
 
 # Initialize Flask app
 app = Flask(__name__)
@@ -155,6 +153,13 @@ config_routes.init_config_routes(
 )
 app.register_blueprint(config_routes.config_bp)
 app.register_blueprint(memory_routes.memory_bp)
+earnings_routes.init_earnings_routes(
+    dashboard_service,
+    state_manager,
+    notification_service,
+    lambda: cached_metrics,
+)
+app.register_blueprint(earnings_routes.earnings_bp)
 
 # Initialize memory manager with required dependencies
 init_memory_manager(
@@ -728,197 +733,6 @@ def format_datetime(value, timezone=None):
         return value
 
 
-# Then update your earnings route
-@app.route("/earnings")
-def earnings():
-    """Serve the earnings page with user's currency and timezone preferences."""
-    try:
-        # Get user's currency and timezone preferences
-        from config import get_currency, get_timezone
-
-        user_currency = get_currency()
-        user_timezone = get_timezone()
-
-        # Define currency symbols for common currencies
-        currency_symbols = {
-            "USD": "$",
-            "EUR": "€",
-            "GBP": "£",
-            "JPY": "¥",
-            "CAD": "C$",
-            "AUD": "A$",
-            "CNY": "¥",
-            "KRW": "₩",
-            "BRL": "R$",
-            "CHF": "Fr",
-        }
-
-        error_message = None
-
-        # Add graceful error handling for earnings data
-        try:
-            # Get earnings data with a longer timeout
-            earnings_data = dashboard_service.get_earnings_data()
-            state_manager.save_last_earnings(earnings_data)
-        except requests.exceptions.ReadTimeout:
-            logging.warning("Timeout fetching earnings data from ocean.xyz - using cached or fallback data")
-            error_message = "Timeout fetching earnings data. Showing cached information."
-            earnings_data = state_manager.get_last_earnings() or {}
-            if not earnings_data:
-                if cached_metrics and "unpaid_earnings" in cached_metrics:
-                    earnings_data = {
-                        "payments": [],
-                        "total_payments": 0,
-                        "total_paid_btc": 0,
-                        "total_paid_sats": 0,
-                        "total_paid_usd": 0,
-                        "unpaid_earnings": cached_metrics.get("unpaid_earnings", 0),
-                        "unpaid_earnings_sats": int(float(cached_metrics.get("unpaid_earnings", 0)) * 100000000),
-                        "est_time_to_payout": cached_metrics.get("est_time_to_payout", "Unknown"),
-                        "monthly_summaries": [],
-                        "timestamp": datetime.now(ZoneInfo(user_timezone)).isoformat(),
-                    }
-                else:
-                    earnings_data = {
-                        "payments": [],
-                        "total_payments": 0,
-                        "total_paid_btc": 0,
-                        "total_paid_sats": 0,
-                        "total_paid_usd": 0,
-                        "unpaid_earnings": 0,
-                        "unpaid_earnings_sats": 0,
-                        "est_time_to_payout": "Unknown",
-                        "monthly_summaries": [],
-                        "timestamp": datetime.now(ZoneInfo(user_timezone)).isoformat(),
-                    }
-
-            notification_service.add_notification(
-                "Data fetch timeout: Unable to fetch payment history data from Ocean.xyz. Showing limited earnings data.",
-                level=NotificationLevel.WARNING,
-                category=NotificationCategory.DATA,
-            )
-        except Exception as e:
-            logging.error(f"Error fetching earnings data: {e}")
-            error_message = f"Error fetching earnings data: {e}"
-            earnings_data = state_manager.get_last_earnings() or {
-                "payments": [],
-                "total_payments": 0,
-                "total_paid_btc": 0,
-                "total_paid_sats": 0,
-                "total_paid_usd": 0,
-                "unpaid_earnings": 0,
-                "unpaid_earnings_sats": 0,
-                "est_time_to_payout": "Unknown",
-                "monthly_summaries": [],
-                "timestamp": datetime.now(ZoneInfo(user_timezone)).isoformat(),
-            }
-
-            notification_service.add_notification(
-                f"Error fetching earnings data: {str(e)}",
-                level=NotificationLevel.ERROR,
-                category=NotificationCategory.DATA,
-            )
-
-        # Convert USD values to user's preferred currency if needed
-        if user_currency != "USD" and earnings_data:
-            # Get exchange rate
-            try:
-                exchange_rates = dashboard_service.fetch_exchange_rates()
-                exchange_rate = exchange_rates.get(user_currency, 1.0)
-
-                # Total paid conversion
-                if "total_paid_usd" in earnings_data:
-                    earnings_data["total_paid_fiat"] = earnings_data["total_paid_usd"] * exchange_rate
-
-                # Monthly summaries conversion
-                if "monthly_summaries" in earnings_data:
-                    for month in earnings_data["monthly_summaries"]:
-                        if "total_usd" in month:
-                            month["total_fiat"] = month["total_usd"] * exchange_rate
-            except Exception as e:
-                logging.error(f"Error converting currency: {e}")
-                # Set fiat values equal to USD as fallback
-                if "total_paid_usd" in earnings_data:
-                    earnings_data["total_paid_fiat"] = earnings_data["total_paid_usd"]
-
-                if "monthly_summaries" in earnings_data:
-                    for month in earnings_data["monthly_summaries"]:
-                        if "total_usd" in month:
-                            month["total_fiat"] = month["total_usd"]
-        else:
-            # If currency is USD, just copy USD values
-            if earnings_data:
-                if "total_paid_usd" in earnings_data:
-                    earnings_data["total_paid_fiat"] = earnings_data["total_paid_usd"]
-
-                if "monthly_summaries" in earnings_data:
-                    for month in earnings_data["monthly_summaries"]:
-                        if "total_usd" in month:
-                            month["total_fiat"] = month["total_usd"]
-
-        return render_template(
-            "earnings.html",
-            earnings=earnings_data,
-            error_message=error_message,
-            user_currency=user_currency,
-            user_timezone=user_timezone,
-            currency_symbols=currency_symbols,
-            current_time=datetime.now(ZoneInfo(user_timezone)).strftime("%b %d, %Y %I:%M:%S %p"),
-        )
-    except Exception as e:
-        logging.error(f"Error rendering earnings page: {e}")
-        import traceback
-
-        logging.error(traceback.format_exc())
-        return render_template(
-            "error.html",
-            message="Failed to load earnings data. Please try again later.",
-        ), 500
-
-
-def payments_to_csv(payments):
-    """Convert a list of payment dictionaries to a CSV string."""
-    with io.StringIO() as output:
-        writer = csv.writer(output)
-        writer.writerow(
-            ["date", "txid", "lightning_txid", "amount_btc", "amount_sats", "status"]
-        )
-        for pay in payments:
-            writer.writerow(
-                [
-                    pay.get("date", ""),
-                    pay.get("txid", ""),
-                    pay.get("lightning_txid", ""),
-                    pay.get("amount_btc", 0),
-                    pay.get("amount_sats", 0),
-                    pay.get("status", ""),
-                ]
-            )
-        return output.getvalue()
-
-
-@app.route("/api/earnings")
-def api_earnings():
-    """API endpoint for earnings data."""
-    try:
-        # Get the earnings data with a reasonable timeout
-        earnings_data = dashboard_service.get_earnings_data()
-        if not isinstance(earnings_data, dict) or earnings_data.get("error"):
-            if isinstance(earnings_data, dict):
-                logging.error("Earnings data error: %s", earnings_data.get("error"))
-            else:
-                logging.error("Earnings data unavailable")
-            return jsonify({"error": "internal server error"}), 500
-        state_manager.save_last_earnings(earnings_data)
-        fmt = request.args.get("format", "json").lower()
-        if fmt == "csv":
-            csv_data = payments_to_csv(earnings_data.get("payments", []))
-            headers = {"Content-Disposition": "attachment; filename=earnings.csv"}
-            return Response(csv_data, mimetype="text/csv", headers=headers)
-        return jsonify(earnings_data)
-    except Exception:
-        logging.exception("Error in earnings API endpoint")
-        return jsonify({"error": "internal server error"}), 500
 
 
 # Add the middleware
